renamed cryptopp/config.h to settings.h
naming an header config.h besides the one generated in the root dir by autoconf is a very bad idea(tm)
