+++ .gitignore
@@ -10,6 +10,7 @@ Makefile.in
 aclocal.m4
 autom4te.cache
 config.guess
+config.h
 config.h.in
 config.log
 config.status
