+++ src/Makefile.am
@@ -6,7 +6,8 @@ bin_PROGRAMS = bitcoind
 bitcoind_SOURCES = util.cpp script.cpp db.cpp net.cpp irc.cpp \
 			main.cpp rpc.cpp init.cpp
 bitcoind_CXXFLAGS = -O2 -Wno-invalid-offsetof -Wformat -I$(DB_CXX_HEADER_PREFIX)
-bitcoind_CPPFLAGS = $(BOOST_CPPFLAGS) $(SSL_CPPFLAGS) $(CRYPTO_CPPFLAGS) $(GTHREAD2_CPPFLAGS)
+bitcoind_CPPFLAGS = -Icryptopp \
+	$(BOOST_CPPFLAGS) $(SSL_CPPFLAGS) $(CRYPTO_CPPFLAGS) $(GTHREAD2_CPPFLAGS)
 bitcoind_LDADD = cryptopp/libcryptopp.la \
    $(COMMON_LIBS) \
    $(BOOST_LIBS) \
