+++ src/cryptopp/Makefile.am
@@ -3,5 +3,5 @@ noinst_LTLIBRARIES = libcryptopp.la
 
 libcryptopp_la_SOURCES = sha.cpp cpu.cpp
 
-noinst_HEADERS = config.h cpu.h cryptlib.h iterhash.h misc.h \
+noinst_HEADERS = settings.h cpu.h cryptlib.h iterhash.h misc.h \
 	  pch.h secblock.h sha.h simple.h smartptr.h stdcpp.h
