+++ src/cryptopp/cpu.h
@@ -10,7 +10,7 @@
 
 #else
 
-#include "config.h"
+#include <settings.h>
 
 #ifdef CRYPTOPP_MSVC6PP_OR_LATER
 	#include <emmintrin.h>
